Fixed not being able to delete immediately after inserting record

+++ app/src/main/java/com/example/paymeback/ui/navigation/PayMeBackNavGraph.kt
@@ -21,7 +21,7 @@ import com.example.paymeback.ui.screens.RecordEditViewModel
 const val HOME_ROUTE: String = "home"
 const val EDIT_PAYMENT_ROUTE: String = "payment_edit"
 const val RECORD_EDIT_ROUTE: String = "record_edit"
-const val DEFAULT_ENTRY_ID: Int = -1
+const val DEFAULT_ENTRY_ID: Long = -1L
 
 @Composable
 fun PayMeBackNavHost(
